+++ types/index.test-d.ts
@@ -1,18 +1,24 @@
+import fastify from 'fastify'
 import healthcheckPlugin, { FastifyHealthcheckOptions } from '..'
 import { expectAssignable } from 'tsd'
-import fastify from 'fastify'
 
 const app = fastify()
-app.register(healthcheckPlugin, {
+const options: FastifyHealthcheckOptions = {
   healthcheckUrl: '/health',
   healthcheckUrlDisable: false,
   healthcheckUrlAlwaysFail: false,
   exposeUptime: false,
   underPressureOptions: { }
+}
+app.register(healthcheckPlugin, options)
+
+app.listen(3000, (err, address) => {
+  if (err) throw err
+  console.log(`Server listening on '${address}' ...`)
 })
 
-expectAssignable < FastifyHealthcheckOptions >({})
-expectAssignable < FastifyHealthcheckOptions >({
+expectAssignable<FastifyHealthcheckOptions>({})
+expectAssignable<FastifyHealthcheckOptions>({
   healthcheckUrl: '/health',
   healthcheckUrlDisable: true,
   healthcheckUrlAlwaysFail: true,